fix warning C4267
We assing a DWORD to wParam when posting WM_DESTROY,
thus we can be sure that there's no loss of data.

+++ src/libs/utils/process_ctrlc_stub.cpp
@@ -108,7 +108,7 @@ int main(int argc, char **)
     while (GetMessage(&msg, NULL, 0, 0))
     {
         if (msg.message == WM_DESTROY)
-            dwExitCode = msg.wParam;
+            dwExitCode = static_cast<DWORD>(msg.wParam);
         TranslateMessage(&msg);
         DispatchMessage(&msg);
     }
